Renaming things to better fit docs.
