+++ docs/index.md
@@ -1,25 +1,28 @@
-# SimpleUAM : Tools for SWRi's UAV and UAM Workflow
-
-SimpleUAM is a set of python libraries and command-line tools for working with
-SWRi's pipeline for UAV development.
-Its main goals are to make the SWRi pipeline easy to deploy and wrap it with
-convenient interfaces for external tools.
+--8<-- "README.md"
 
 ## Organization
 
-SimpleUAM organizes its major components into nodes, each of which perform
-some basic task:
+<figure markdown>
+  ![Project Components](assets/component-diagram.png)
+  <figcaption>SimpleUAM Component Structure</figcaption>
+</figure>
 
-  - **Clients**: Make requests for UAV/UAM design analysis.
-  - **Message Broker**: Distribute analysis requests from clients to available
+SimpleUAM organizes itself into components, each of which perform
+some basic task needed to evaluate UAV or UAM designs:
+
+  - **Client Nodes**: Makes requests for UAV/UAM design analysis.
+  - **Message Broker**: Distributes analysis requests from clients to available
     workers.
-  - **Workers**: Analyze designs with SWRi's pipelines.
-  - **Engineering Corpus**: Provide component and design data for use during analysis.
-  - **License Servers**: Provide floating Creo licenses to workers.
-  - **Results Storage**: Store the results of design analyses.
+  - **Worker Nodes**: Analyzes designs with SWRi's pipelines, placing the results
+    into storage.
+  - **Engineering Corpus**: Provides component and design data to the
+    worker nodes during analysis.
+  - **License Servers**: Provides floating Creo licenses to workers.
+  - **Results Storage**: Store the results of design analyses on the file system,
+    whether it's in a local folder or a network drive.
 
-These nodes aren't tied to specific machines and can all happily coexist on a
-single computer or support multiple servers cooperating on each task.
+These components aren't tied to specific machines and can all happily coexist on
+a single computer or support multiple servers cooperating on each task.
 
 This project provides support for versions of all of these nodes including
 setup scripts and python libraries for interacting with them.
@@ -27,3 +30,18 @@ setup scripts and python libraries for interacting with them.
 ## Links
 
  - [**Install Instructions**](setup/intro.md)
+ - [**Configuration Instructions**](usage/config.md)
+
+## Repo Organization
+
+```
+<repo-root>
+├── LICENSE           # License File
+├── README.md         # Readme File
+├── pyproject.toml    # Project Setup Info. For use with pdm, poetry, or similar.
+├── mkdocs.yml        # Documentation Configuration via mkdocs
+│
+├── config/   # Config files for type checking, linting, etc...
+├── docs/     # Source for gh-pages documentation.
+└── src/    # Python source root
+```
