+++ docs/setup/aws-instance.md
@@ -2,12 +2,12 @@
 
 This section of the guide will cover:
 
-  - Creating the various EC2 instances needed for each node or combination of
-    nodes.
+  - Creating the various EC2 instances needed for each component or combination of
+    component.
 
 ## Provisioning an Instance
 
-Setup an EC2 instance for running various nodes.
+Setup an EC2 instance for running various component.
 This only covers the various windows nodes that SimpleUAM currently supports
 and needs to be repeated for each node.
 
@@ -57,35 +57,38 @@ and needs to be repeated for each node.
         - **Advanced Details:**
             - **Elastic GPU:** eg1.medium
 
-    ??? note "Corpus Server Minimum Requirements"
+        This can get CPU bound, so it might make sense to bump it up.
 
-        The corpus server provided with SimpleUAM is a minimal stub that
-        can run on a worker node alongside a worker.
+    ??? note "Corpus DB Minimum Requirements"
 
-        A proper corpus server should probably be run on a linux instance
-        in this VPC but setting that up is beyond the current scope of
-        this library.
+        **Option 1:** If you need a read-only corpus DB itermittently for
+        generating a static corpus, we reccomend just running the stub DB on a
+        worker temporarily.
 
-        If you must run the stub server is a persistant manner, then the
-        worker node minimum requirements are a decent starting point. Maybe
-        bump up the CPU a bit since it does tend to get saturated when the
-        stub is being queried a lot.
+        **Option 2:** If you need a corpus DB for an alternate analysis
+        pipeline and are fine with a non-persistent corpus then use a
+        stub DB on a worker.
+
+        **Option 3:** If you need a corpus that is persistent, shared between
+        multiple workers, or needs to be performant then you should set up
+        janus-graph on a linux machine.
+
+        This is outside the scope of this guide and you should try to follow
+        SWRI's instructions.
 
     ??? note "Message Broker Minimum Requirements"
 
-        We haven't tested the message broker enough to have good minimum
-        requirements, but it can ride sidecar with a worker node.
+        **Option 1:** If you want a non-backendless broker or one that's not
+        wasting your money set up a linux instance and skip to the section
+        on [message broker setup](broker.md). I'm not sure what minimum
+        requirements make sense.
 
-        That said running a broker on windows for more than development work is
-        not a great idea.
-        Both supported brokers, rabbitmq and redis, are used out of the box
-        with default settings.
-        They both have much stronger linux support than windows support and
-        they'll be more performant when running on an OS that's actually good.
+        **Option 2:** If you want a backendless message broker for intermittent
+        testing or development work then install it sidecar with a worker node.
 
-        If you're running a broker in production just stick
-        [rabbitmq](https://www.rabbitmq.com/download.html) or
-        [redis](https://redis.io/download/) on linux instance.
+        **Option 3:** If you want a backendless broker and are fine with lower
+        performance the specifications for a worker node on a windows machine
+        be fine.
 
 - Find the instance you just created with "Name" `<instance.name>`.
     - Save the "Instance ID" as: `<instance.id>`
